# perf: Use mmap-ed memory if possible in Parquet reader